Max script file/block size from 8KB -> 32KB
Changed `config_loader` so that it dynamically allows for file sizes up to 32KB.
Behavior: The `ConfigLoader` reads up to 32KB before quitting. If at any point in that 32KB range there is an `@def`, another 32KB is able to be read.
If the file has no `@def` and it's less than or eq to 32KB, the file is considered passing by the loader.
Adjusted how `ConfigLoader` consumes `@def` to where it properly aligns with it's relative position in an actual region.

NOTE: Any tooling that uses this now needs to use the region metadata which tracks the exact line position it starts at and bring the line to it's actual absolute position, rather than the line relative to the region itself.

+++ chrn_core/compilation/src/parser.rs
@@ -42,7 +42,10 @@ pub fn parse(
     let mut ast_info = AstInfo::new();
 
     let mut state = ParserState::new();
-    let budget = ParserBudget::new(chrn_utils::MAX_RECURSIVE_DEPTH, chrn_utils::MAX_EXPR_NODES);
+    let budget = ParserBudget::new(
+        chrn_utils::MAX_RECURSIVE_DEPTH,
+        chrn_utils::MAX_EXPR_NODES as usize,
+    );
     let mut ctx = ParserContext::new(cfg, region, tokens);
 
     // Skipping possible @def first since it is recognized as it's own token
@@ -1213,6 +1216,7 @@ fn parse_type_expr(
     interner: &Intern,
 ) -> Result<SpannedContainer<TypeExpr>, Token> {
     // SAFETY:
+    //TODO: Foot
     let _guard = budget.increase_depth().map_err(|_| {
         let msg = format!(
             "Reached max recursive depth of {}",
